[charts] Organize series config
