+++ packages/x-charts/src/BarChart/seriesConfig/bar/getSeriesWithDefaultValues.ts
@@ -1,4 +1,4 @@
-import type { GetSeriesWithDefaultValues } from '../../internals/plugins/models/seriesConfig';
+import type { GetSeriesWithDefaultValues } from '../../../internals/plugins/models/seriesConfig';
 
 const getSeriesWithDefaultValues: GetSeriesWithDefaultValues<'bar'> = (
   seriesData,
