+++ packages/x-charts/src/BarChart/seriesConfig/bar/legend.ts
@@ -1,6 +1,6 @@
-import type { LegendItemParams } from '../../ChartsLegend';
-import { getLabel } from '../../internals/getLabel';
-import { LegendGetter } from '../../internals/plugins/models';
+import type { LegendItemParams } from '../../../ChartsLegend';
+import { getLabel } from '../../../internals/getLabel';
+import { LegendGetter } from '../../../internals/plugins/models';
 
 const legendGetter: LegendGetter<'bar'> = (params) => {
   const { seriesOrder, series } = params;
