+++ packages/x-charts/src/BarChart/seriesConfig/bar/tooltip.ts
@@ -1,5 +1,5 @@
-import { getLabel } from '../../internals/getLabel';
-import type { AxisTooltipGetter, TooltipGetter } from '../../internals/plugins/models';
+import { getLabel } from '../../../internals/getLabel';
+import type { AxisTooltipGetter, TooltipGetter } from '../../../internals/plugins/models';
 
 const tooltipGetter: TooltipGetter<'bar'> = (params) => {
   const { series, getColor, identifier } = params;
