+++ packages/x-charts/src/LineChart/seriesConfig/index.ts
@@ -7,7 +7,7 @@ import tooltipGetter, { axisTooltipGetter } from './tooltip';
 import getSeriesWithDefaultValues from './getSeriesWithDefaultValues';
 import tooltipItemPositionGetter from './tooltipPosition';
 
-export const seriesConfig: ChartSeriesTypeConfig<'line'> = {
+export const lineSeriesConfig: ChartSeriesTypeConfig<'line'> = {
   colorProcessor: getColor,
   seriesProcessor,
   legendGetter,
